requirements.txt
# ───────────────────────── core ─────────────────────────
nest_asyncio>=1.6.0
aiohttp>=3.11          # 3.11.18 on today’s index
requests>=2.32.0

# ────────────────────── telegram stack ──────────────────
python-telegram-bot==20.3        # wants httpx ~=0.24
httpx==0.24.1                    # 0.24.1 → httpcore 0.16.*

# ───────────────────── system monitoring ────────────────
psutil>=5.9.0

# ──────────────────── solana tool-chain ─────────────────
# Matching pair – DO NOT bump one without the other
solana==0.35.0        # requires   solders >=0.21,<0.22
solders==0.21.0
base58>=2.1.0

# ───────────────── security / crypto ────────────────────
pycryptodome>=3.22.0
cryptography>=44.0.0

# ────────────────────── utilities ───────────────────────
tenacity>=8.2.0
asyncio-throttle>=1.0.2
ujson>=5.10.0
python-dotenv>=1.0.1
backoff>=2.2.1
python-json-logger>=2.0.7

# ───────────────────── data / ML ────────────────────────
numpy>=1.24,<2.3
pandas>=2.0,<3.0
matplotlib>=3.7.0
